refactor core app into app.py
