+++ dinghy_ping/services/api.py
@@ -1,20 +1,17 @@
-import responder
 import requests
 from prometheus_client import Counter, Summary, start_http_server
 import time
 import asyncio
-import os
 import json
-import sys
-sys.path.insert(0, './dinghy_ping/models/')
-import data
-import dinghy_dns
+import dinghy_ping.models.data as data
+import dinghy_ping.models.dinghy_dns as dinghy_dns
 import dns.rdatatype
 import socket
 import logging
 from urllib.parse import urlparse
 from kubernetes import client, config
 from kubernetes.client.rest import ApiException
+from dinghy_ping.app import api, redis_host, k8s_client
 
 # Prometheus metrics
 COMPLETED_REQUEST_COUNTER = Counter('dingy_pings_completed', 'Count of completed dinghy ping requests')
@@ -22,24 +19,6 @@ FAILED_REQUEST_COUNTER = Counter('dingy_pings_failed', 'Count of failed dinghy p
 REQUEST_TIME = Summary('dinghy_request_processing_seconds', 'Time spent processing request')
 TAIL_LINES_DEFAULT = 100
 LOGS_PREVIEW_LENGTH = 1000
-TEMPLATE_DIR = 'dinghy_ping/views/templates/'
-
-# Configure kubernetes client
-if not "IN_TRAVIS" in os.environ:
-    config.load_incluster_config()
-    k8s_client = client.CoreV1Api()
-
-def to_pretty_json(value):
-    return json.dumps(value, sort_keys=True,
-                      indent=4, separators=(',', ': '))
-
-print(os.getcwd())
-api = responder.API(title="Dinghy Ping", templates_dir=TEMPLATE_DIR, version="1.0", openapi="3.0.0", docs_route="/docs")
-api.jinja_env.filters['tojson_pretty'] = to_pretty_json
-
-# For local mac docker image creation and testing, switch to host.docker.internal
-redis_host = os.getenv("REDIS_HOST", default="127.0.0.1")
-
 
 @api.route("/")
 def dinghy_html(req, resp):
